+++ CumulusMX/Lang.cs
@@ -200,5 +200,8 @@
 		public string WebTagRecTimeDate { get; set; }
 		public string WebTagRecDate { get; set; }
 		public string WebTagRecDryWetDate { get; set; }
+		// Snow
+		public string SnowDepth { get; set; }
+		public string Snow24h { get; set; }
 	}
 }