+++ CumulusMX/WS2300Station.cs
@@ -218,6 +218,7 @@ namespace CumulusMX
 			bool rolloverdone = luhour == rollHour;
 
 			bool midnightraindone = luhour == 0;
+			bool rollover9amdone = luhour == 9;
 
 			double prevraintotal = -1;
 			double raindiff, rainrate;
@@ -265,6 +266,12 @@ namespace CumulusMX
 					midnightraindone = true;
 				}
 
+				// 9am rollover items
+				if (h == 9 && !rollover9amdone)
+				{
+					Reset9amTemperatures(timestamp);
+					rollover9amdone = true;
+				}
 				// Humidity ====================================================================
 				if ((historydata.inHum > 0) && (historydata.inHum <= 100))
 					DoIndoorHumidity(historydata.inHum);